Refactor: Header 컴포넌트에서 NavBar 컴포넌트 분리
