+++ src/components/NavBar/styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const styles = {
+  navList: css({
+    display: 'flex',
+  }),
+};
